Move logic to classify from dialog to app
